foundry_wallets::wallet_browser

Module types

Structs

BrowserSignRequest

Represents a signing request sent to the browser wallet.

BrowserSignResponse 🔒

Represents a signing response sent from the browser wallet.

BrowserSignTypedDataRequest

Represents a typed data signing request sent to the browser wallet.

BrowserTransactionRequest

Represents a transaction request sent to the browser wallet.

BrowserTransactionResponse 🔒

Represents a transaction response sent from the browser wallet.

Connection

Represents an active connection to a browser wallet.

SignRequest

Enums

BrowserApiResponse 🔒

Response format for API endpoints.

SignType